+++ src/cursor/cursor_row.rs
@@ -42,14 +42,14 @@ impl<'a> std::fmt::Debug for CursorRow<'a> {
 impl<'a> CursorRow<'a> {
     /// Get the resource ID from the arguments buffer which dynamically changes
     /// after each cursor advance.
-    fn resource_id(&self, term_index: usize) -> Result<Option<u64>, RDFStoreError> {
+    fn resource_id(&self, term_index: u64) -> Result<Option<u64>, RDFStoreError> {
         self.opened.resource_id(term_index)
     }
 
     /// Returns the resource bound to the given index in the current answer row.
     fn lexical_value_with_id(&self, resource_id: u64) -> Result<Option<Literal>, RDFStoreError> {
         let mut buffer = [0u8; 102400]; // TODO: Make this dependent on returned info about buffer size too small
-        let mut lexical_form_size = 0 as usize;
+        let mut lexical_form_size = 0_u64;
         let mut datatype_id: u8 = DataType::UnboundValue as u8;
         let mut resource_resolved = false;
         tracing::trace!("CCursor_getResourceLexicalForm({resource_id}):");
@@ -59,7 +59,7 @@ impl<'a> CursorRow<'a> {
                 self.opened.cursor.inner,
                 resource_id,
                 buffer.as_mut_ptr() as *mut i8,
-                buffer.len(),
+                buffer.len() as u64,
                 &mut lexical_form_size,
                 &mut datatype_id as *mut u8,
                 &mut resource_resolved,
@@ -84,7 +84,7 @@ impl<'a> CursorRow<'a> {
 
     /// Get the value in lexical form of a term in the current solution /
     /// current row with the given term index.
-    pub fn lexical_value(&self, term_index: usize) -> Result<Option<Literal>, RDFStoreError> {
+    pub fn lexical_value(&self, term_index: u64) -> Result<Option<Literal>, RDFStoreError> {
         let resource_id = self.resource_id(term_index)?;
         if event_enabled!(tracing::Level::TRACE) {
             tracing::trace!(
